Separate Projects data from index
