+++ projectsdata.js
@@ -1,0 +1,34 @@
+const projectsData = [
+  {
+    title: 'To do List',
+    detailsTitle: 'Self',
+    detailsText: 'Front End Dev',
+    detailsYear: '2023',
+    description: 'A minimalistic to-do list app to keep track of your tasks, explore the app to know more!',
+    descriptionLong: 'With a To Do list you can manage your daily tasks. It is simple to use in a way that you can add, delete and mark as done your tasks. You can also filter your tasks by status. It is a great way to keep track of your tasks and be more productive.',
+    imageUrl: 'Assets/portfolioImages/Project To Do List.png',
+    tags: ['HTML', 'CSS', 'Javascript'],
+  },
+  {
+    title: 'Pokémon APP',
+    detailsTitle: 'React',
+    detailsText: 'Full Stack Dev',
+    detailsYear: '2023',
+    description: 'An app to search for your favorite pokémon, explore the app to know more!',
+    descriptionLong: 'This app is a great way to search for your favorite pokémon. You can interact with other users by commenting on their posts. Share your houghts about your favorite pokémon now.',
+    imageUrl: 'Assets/portfolioImages/Poke 1 - wide.png',
+    tags: ['React', 'Redux', 'Javascript'],
+  },
+  {
+    title: 'Space Travelers Hub',
+    detailsTitle: 'React',
+    detailsText: 'Full Stack Dev',
+    detailsYear: '2023',
+    description: 'An app to book your next space travel, explore the app to know more!',
+    descriptionLong: "Space Travelers' Hub is a company that provides commercial and scientific space travel services. They also allow rockets and capsules to be booked. This app is a great way to book your next space travel. You can interact with other users by booking rockets and capsules.",
+    imageUrl: 'Assets/portfolioImages/space rockets.png',
+    tags: ['React', 'Redux', 'Javascript'],
+  },
+];
+
+export default projectsData;
